fi: fix the mood entry issue, fix the emoji on the page details, headers space, and protected root after login
